Fix kernel module build warnings

+++ tracing_library/source/tracepoint.c
@@ -46,8 +46,8 @@ void _clltk_static_tracepoint_with_args(_clltk_tracebuffer_handler_t *handler,
 										...)
 {
 	if (unlikely(false == _CLLTK_FILE_OFFSET_IS_STATIC(in_file_offset))) {
-		ERROR_LOG("invalid in_file_offset(%lu) at %s:%d for %s", in_file_offset, file, line,
-				  handler->definition.name);
+		ERROR_LOG("invalid in_file_offset(%llu) at %s:%d for %s",
+				  (unsigned long long)in_file_offset, file, line, handler->definition.name);
 		return;
 	}
 
@@ -111,8 +111,8 @@ void _clltk_static_tracepoint_with_dump(_clltk_tracebuffer_handler_t *handler,
 										const void *address, uint32_t size_in_bytes)
 {
 	if (unlikely(false == _CLLTK_FILE_OFFSET_IS_STATIC(in_file_offset))) {
-		ERROR_LOG("invalid in_file_offset(%lu) at %s:%d for %s", in_file_offset, file, line,
-				  handler->definition.name);
+		ERROR_LOG("invalid in_file_offset(%llu) at %s:%d for %s",
+				  (unsigned long long)in_file_offset, file, line, handler->definition.name);
 		return;
 	}
